# Imporved performance of Utf16ValueStringBuilder.AppendFormat on Slow Span Environment