For environments like Mono and .NET Framework, I reduced the Span used in Utaf16ValueStringBuilder.AppendFormat.

# Benchmark on .NET Framework

```
$ dotnet run -c Release --framework net47
Available Benchmarks:
  #0 AppendPerformance
  #1 FormatBenchmark
...
1
```

```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19043
AMD Ryzen 5 5600X, 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=6.0.100-preview.4.21255.9
  [Host]   : .NET Core 5.0.6 (CoreCLR 5.0.621.22011, CoreFX 5.0.621.22011), X64 RyuJIT  [AttachedDebugger]
  ShortRun : .NET Framework 4.8 (4.8.4341.0), X64 RyuJIT

Job=ShortRun  Runtime=.NET 4.7  IterationCount=1
LaunchCount=1  WarmupCount=1
```

## Before
|                Method |      Mean | Error | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |----------:|------:|------:|-------:|------:|------:|----------:|
|          StringFormat | 155.02 ns |    NA |  1.00 | 0.1013 |     - |     - |     169 B |
|         ZStringFormat | 149.62 ns |    NA |  0.97 | 0.0336 |     - |     - |      56 B |
| ZStringPreparedFormat | 122.82 ns |    NA |  0.79 | 0.0336 |     - |     - |      56 B |
| StringFormatterFormat |  81.76 ns |    NA |  0.53 | 0.0337 |     - |     - |      56 B |
## After

|                Method |      Mean | Error | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |----------:|------:|------:|-------:|------:|------:|----------:|
|          StringFormat | 153.41 ns |    NA |  1.00 | 0.1013 |     - |     - |     169 B |
|         ZStringFormat | 133.61 ns |    NA |  0.87 | 0.0336 |     - |     - |      56 B |
| ZStringPreparedFormat | 122.32 ns |    NA |  0.80 | 0.0336 |     - |     - |      56 B |
| StringFormatterFormat |  81.04 ns |    NA |  0.53 | 0.0337 |     - |     - |      56 B |